refactor: moving item parsers
