+++ test/item.weapon.test.ts
@@ -1,4 +1,5 @@
-import { buildDamageParts, getRange, parseSpellCone, parseSpellSphere } from '../src/module/item/weapon';
+import { buildDamageParts } from '../src/module/item/weapon';
+import { parseRange, parseSpellCone, parseSpellSphere } from '../src/module/item/parsers';
 
 describe('buildDamage', () => {
   it('should build damage from a melee weapon', () => {
@@ -37,11 +38,11 @@ describe('parseSpellSphere', () => {
   });
 });
 
-describe('getRange', () => {
+describe('parseRange', () => {
   it('should get the range of a spell', () => {
     const text =
       'Poison gas fills a 20-foot-radius sphere centered on a point Big Bara can see within 50 feet of her. The gas spreads around corners and remains until the start of Big Bara’s next turn. Each creature that starts its turn in the gas must succeed on a DC 16 Constitution saving throw or be poisoned for 1 minute. A creature can repeat the saving throw at the end of each of its turns, ending the effect on itself on a success.';
-    const range = getRange(text);
+    const range = parseRange(text);
     expect(range).toBeDefined();
     expect(range?.value).toEqual(50);
   });
